+++ package.json
@@ -16,6 +16,7 @@
         "bcrypt-ts": "^7.1.0",
         "class-variance-authority": "^0.7.1",
         "clsx": "^2.1.1",
+        "jose": "^6.0.12",
         "lucide-react": "^0.525.0",
         "mysql2": "^3.14.2",
         "next": "15.3.5",
