+++ src/app/api/login/route.ts
@@ -1,8 +1,10 @@
-import { authenticateUser } from "@/libs/auth";
-import { errorHandler } from "@/libs/errors/errorHandler";
+import { authenticateUser, generateToken } from "@/libs/server/auth";
+import { errorHandler } from "@/libs/server/errors/errorHandler";
 import { loginSchema } from "@/libs/validation/authSchema";
+import { cookies } from "next/headers";
+import { NextRequest, NextResponse } from "next/server";
 
-export async function POST(req: Request) {
+export async function POST(req: NextRequest) {
     const body = await req.json();
     try {
         const validationResult = loginSchema.safeParse(body);
@@ -11,8 +13,20 @@ export async function POST(req: Request) {
         }
         const { email, password } = validationResult.data;
         const user = await authenticateUser(email, password);
-        // Générer JWT
-        return new Response(JSON.stringify({ message: "Connexion réussie" }), { status: 200 });
+        const jwt = await generateToken({ id: user.id, email: user.email });
+
+        const response = NextResponse.json({ message: "Connexion réussie" }, { status: 200 });
+        response.cookies.set({
+            name: "auth_token",
+            value: jwt,
+            httpOnly: true,
+            path: "/",
+            secure: process.env.NODE_ENV === "production",
+            sameSite: "lax",
+            maxAge: 24 * 60 * 60,
+        });
+
+        return response;
     } catch (error) {
         console.log(error);
         return errorHandler(error);
