+++ src/libs/server/errors/customErrors.ts
